Made it work with BCB6 Some upper/lower case filename conversions Some cosmetics to the source

+++ AREAFIX.C
@@ -449,7 +449,7 @@ s16 areaFix(internalMsgType *message)
   s32             msgNum1, msgNum2;
   rawEchoType     rawEchoInfo2;
   u16             areaCount;
-  u16             areaFixCount = 0;
+  int             areaFixCount = 0;
   u16             activeAreasCount = 0;
   areaFixListType *areaFixList;
   areaFixType     areaFixRec;
@@ -1676,8 +1676,9 @@ Send:
     {
       if ((helpHandle = openP(tempStr, O_RDWR|O_BINARY|O_CREAT|O_TRUNC|O_DENYNONE, S_IREAD|S_IWRITE)) != -1)
       {
-        if (write (helpHandle, areaFixList,
-                   areaFixCount * sizeof(areaFixType)) == areaFixCount * sizeof(areaFixType))
+        if (  write(helpHandle, areaFixList, areaFixCount * sizeof(areaFixType))
+           == (int)(areaFixCount * sizeof(areaFixType))
+           )
         {
           lseek (helpHandle, 0, SEEK_SET);
           while (read(helpHandle, &areaFixRec, sizeof(areaFixType)) > 0)
